Issue #3030929 by kiwimind, Kavichandran P, c.nish2k3, quietone: User migrations tests in wrong directory
(cherry picked from commit bba5f2053a7ee309a4da6aa23d486f44ce8939e2)
