+++ modules/user/tests/src/Kernel/Migrate/d6/MigrateUserProfileEntityFormDisplayTest.php
@@ -1,16 +1,16 @@
 <?php
 
-namespace Drupal\Tests\user\Kernel\Migrate;
+namespace Drupal\Tests\user\Kernel\Migrate\d6;
 
-use Drupal\Core\Entity\Entity\EntityViewDisplay;
+use Drupal\Core\Entity\Entity\EntityFormDisplay;
 use Drupal\Tests\migrate_drupal\Kernel\d6\MigrateDrupal6TestBase;
 
 /**
- * Tests the user profile entity display migration.
+ * Tests the user profile entity form display migration.
  *
  * @group migrate_drupal_6
  */
-class MigrateUserProfileEntityDisplayTest extends MigrateDrupal6TestBase {
+class MigrateUserProfileEntityFormDisplayTest extends MigrateDrupal6TestBase {
 
   /**
    * {@inheritdoc}
@@ -20,23 +20,23 @@ class MigrateUserProfileEntityDisplayTest extends MigrateDrupal6TestBase {
     $this->executeMigrations([
       'user_profile_field',
       'user_profile_field_instance',
-      'user_profile_entity_display',
+      'user_profile_entity_form_display',
     ]);
   }
 
   /**
    * Tests migration of user profile fields.
    */
-  public function testUserProfileFields() {
-    $display = EntityViewDisplay::load('user.user.default');
+  public function testUserProfileEntityFormDisplay() {
+    $display = EntityFormDisplay::load('user.user.default');
 
     // Test a text field.
     $component = $display->getComponent('profile_color');
-    $this->assertIdentical('text_default', $component['type']);
+    $this->assertIdentical('text_textfield', $component['type']);
 
     // Test a list field.
     $component = $display->getComponent('profile_bands');
-    $this->assertIdentical('text_default', $component['type']);
+    $this->assertIdentical('text_textfield', $component['type']);
 
     // Test a date field.
     $component = $display->getComponent('profile_birthdate');
@@ -48,9 +48,10 @@ class MigrateUserProfileEntityDisplayTest extends MigrateDrupal6TestBase {
     // Test PROFILE_HIDDEN field is hidden.
     $this->assertNull($display->getComponent('profile_sold_to'));
 
-    // Test a checkbox field.
+    // Test that a checkbox field has the proper display label setting.
     $component = $display->getComponent('profile_really_really_love_mig');
-    $this->assertIdentical('list_default', $component['type']);
+    $this->assertIdentical('boolean_checkbox', $component['type']);
+    $this->assertIdentical(TRUE, $component['settings']['display_label']);
   }
 
 }
